requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
audioop-lts==0.2.1
cfgv==3.4.0
discord-protos==0.0.2
discord.py==2.4.0
distlib==0.3.9
filelock==3.16.1
frozenlist==1.4.1
identify==2.6.1
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
nodeenv==1.9.1
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
propcache==0.2.0
protobuf==5.28.2
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dotenv==1.0.1
PyYAML==6.0.2
typing_extensions==4.12.2
virtualenv==20.27.0
yarl==1.15.5
